[FE] fix: merge 과정에서 생긴 오류들 수정
- 중복되는 코드 삭제
- style 폴더 삭제
- import 경로 수정
- 몇몇 컴포넌트 prop type 추가
- Provider들을 index.tsx로 이동
