+++ WEB/FE/.eslintrc.json
@@ -28,14 +28,6 @@
       "jsx": true
     }
   },
-  "settings": { 
-    "import/parsers": {
-      "@typescript-eslint/parser": [".ts", ".tsx"]
-    },
-    "import/resolver": {
-       "typescript": "./tsconifg.json"
-    }
-  },
   "rules": {
     "linebreak-style": 0,
     "import/no-dynamic-require": 0,
@@ -52,7 +44,6 @@
     "@typescript-eslint/no-empty-interface": 0,
     "@typescript-eslint/no-explicit-any": 0,
     "no-shadow": "off",
-    // "@typescript-eslint/no-shadow": ["error"],
     "react/prop-types": 0
   },
   "settings": { 
@@ -62,5 +53,5 @@
     "import/resolver": { 
       "typescript": "./tsconfig.json" 
     } 
-  },
+  }
 }
