+++ WEB/FE/src/components/common/Button.tsx
@@ -1,22 +1,17 @@
 import React from 'react';
 import styled from 'styled-components';
-import { Colors } from '@styles/variable';
 
 const ButtonTag = styled.button`
   height: 36px;
   border: none;
-  background-color: ${Colors.GrayBrown};
-  color : ${Colors.White};
+  background-color: ${({ theme }) => theme.color.GrayBrown};
+  color: ${({ theme }) => theme.color.White};
   width: 100%;
 `;
 
-const Button = (props: any) => { 
+const Button = (props: any) => {
   const { name, buttonEvent } = props;
-  return (
-    <>
-      <ButtonTag onClick={(e) => buttonEvent(e)}>{name}</ButtonTag>
-    </>
-  );
+  return <ButtonTag onClick={(e) => buttonEvent(e)}>{name}</ButtonTag>;
 };
 
 export default Button;
