+++ WEB/FE/src/pages/SignUpPage.tsx
@@ -1,12 +1,8 @@
 import React from 'react';
-import SignUpComponent from '../components/SignUpComponent/SignUpComponent';
+import SignUpComponent from '@components/SignUpComponent/SignUpComponent';
 
 const SignUpPage = () => {
-  return (
-    <div>
-      <SignUpComponent />
-    </div>
-  );
+  return <SignUpComponent />;
 };
 
 export default SignUpPage;
