--- WEB/FE/src/style/theme.ts
@@ -1,22 +1,0 @@
-enum Colors {
-  GrayBrown = '#5C4D43',
-  Puple = '#3F3D56',
-  Mustard = '#C57A07',
-  Black = '#333333',
-  Gray = '#D9D8DD',
-  White = '#FFFFFF',
-  border = '#eaecef',
-}
-
-enum FontSize {
-  xl = '32px',
-  lg = '20px',
-  md = '16px',
-  sm = '14px',
-  xs = '12px',
-}
-
-export const theme = {
-  color: Colors,
-  fontSize: FontSize,
-};
