+++ WEB/FE/src/styles/theme.ts
@@ -1,16 +1,22 @@
-export enum Colors {
+enum Colors {
   GrayBrown = '#5C4D43',
   Puple = '#3F3D56',
   Mustard = '#C57A07',
   Black = '#333333',
   Gray = '#D9D8DD',
   White = '#FFFFFF',
+  border = '#eaecef',
 }
 
-export enum FontSize {
+enum FontSize {
   xl = '32px',
   lg = '20px',
   md = '16px',
   sm = '14px',
   xs = '12px',
 }
+
+export const theme = {
+  color: Colors,
+  fontSize: FontSize,
+};
